# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LMS API",
    "version": "2.0.0",
    "description": "Lead Management System API. Internal management routes require a Bearer token obtained from POST /v2/auth/login (Cognito-backed). External lead intake routes are open and authenticate via campaign_id + campaign_key in the request body."
  },
  "servers": [
    {
      "url": "https://3ifu8b0q2h.execute-api.us-east-1.amazonaws.com/dev/v2",
      "description": "Development - Internal API (Bearer token)"
    },
    {
      "url": "https://uj580pu31h.execute-api.us-east-1.amazonaws.com/dev/v2",
      "description": "Development - External Leads API (open \u2014 campaign_id + campaign_key auth)"
    }
  ],
...
  "paths": {
...
    "/campaigns": {
      "post": {
        "summary": "Create campaign (starts as DRAFT)",
        "tags": ["campaigns"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignRequest"
              },
              "examples": {
                "default": {
                  "value": {
                    "name": "Spring Promo"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Campaign created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        }
      },
      "get": {
        "summary": "List campaigns",
        "tags": ["campaigns"],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/CampaignStatus"
            }
          },
          {
            "name": "includeDeleted",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "When true, soft-deleted campaigns are included in the response"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "lastEvaluatedKey",
            "in": "query",
            "schema": {
              "type": "string",
              "description": "Base64-encoded pagination token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/campaigns/{id}": {
      "get": {
        "summary": "Get campaign by id",
        "tags": ["campaigns"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CampaignId"
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "summary": "Update campaign (name only)",
        "tags": ["campaigns"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CampaignId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampaignRequest"
              },
              "examples": {
                "default": {
                  "value": {
                    "name": "Spring Promo (v2)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Campaign updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "summary": "Delete campaign (soft by default; ?permanent=true for hard delete)",
        "description": "Deletion allowed only when status is DRAFT or TEST, with no linked clients, affiliates, or leads. Hard delete is allowed only if the campaign has never had participants or leads. Without `?permanent=true`: soft-deletes the campaign (sets `is_deleted=true`, `active=false`, `deleted_by`, `deleted_at`).",
        "tags": ["campaigns"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CampaignId"
          },
          {
            "name": "permanent",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "When true, permanently removes the campaign and all participant links."
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted (soft or permanent)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/campaigns/{id}/clients": {
      "post": {
        "summary": "Link client to campaign (always linked as TEST; update status via participant endpoint)",
        "tags": ["campaigns"],
        "parameters": [
          {
            "$ref": "#/components/parameters/CampaignId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkClientRequest"
              },
              "examples": {
                "default": {
                  "value": {
                    "client_id": "CLABCDEFG"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Linked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignResponse"
                }
              }
            }
          },
          "404": {
            "description": "Campaign not found"
          }
        }
      }
    },
    "/campaigns/{id}/clients/{clientId}": {
      "put": {
        "summary": "Update linked client campaign_status (TEST/LIVE/DISABLED)",
        "tags": ["campaigns"],
        "parameters": [
          {
            "$ref": "#/components/parameters/CampaignId"
          },
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateParticipantStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Client status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "summary": "Remove linked client from campaign",
        "description": "Removal is blocked if the campaign has leads; disable the client instead.",
        "tags": ["campaigns"],
        "parameters": [
          {
            "$ref": "#/components/parameters/CampaignId"
          },
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT id_token obtained from POST /v2/auth/login. Include as Authorization: Bearer <id_token> on all internal API requests. Note: the API Gateway Cognito authorizer validates the id_token, not the access_token."
      }
    },
    "schemas": {
...
      "CampaignStatus": {
        "type": "string",
        "enum": ["DRAFT", "TEST", "ACTIVE", "INACTIVE"],
        "example": "TEST"
      },
...
      "CreateCampaignRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "description": "Campaigns are created as DRAFT by default."
      },
      "UpdateCampaignRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "description": "Allows updating the campaign name."
      },
...
      "CampaignResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Campaign"
          }
        }
      },
      "CampaignListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "count": {
            "type": "integer"
          },
          "lastEvaluatedKey": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Campaign"
            }
          }
        }
      },
...
    },
    "parameters": {
...
      "CampaignId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "example": "CMABC12345"
      },
...
    }
  }
}